+++ shuup/admin/modules/sales_units/views/list.py
@@ -11,19 +11,26 @@ from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.utils.picotable import Column, TextFilter
 from shuup.admin.utils.views import PicotableListView
-from shuup.core.models import SalesUnit
+from shuup.core.models import DisplayUnit, SalesUnit
 
 
-class SalesUnitListView(PicotableListView):
-    model = SalesUnit
+class UnitListView(PicotableListView):
     default_columns = [
         Column("name", _(u"Name"), sort_field="translations__name", display="name", filter_config=TextFilter(
             filter_field="translations__name",
             placeholder=_("Filter by name...")
         )),
-        Column("short_name", _(u"Short Name"), sort_field="translations__short_name", display="short_name"),
+        Column("symbol", _(u"Symbol"), sort_field="translations__symbol", display="symbol"),
         Column("decimals", _(u"Allowed decimals")),
     ]
 
     def get_queryset(self):
-        return SalesUnit.objects.all()
+        return self.model.objects.all()
+
+
+class SalesUnitListView(UnitListView):
+    model = SalesUnit
+
+
+class DisplayUnitListView(UnitListView):
+    model = DisplayUnit