+++ shuup/core/models/__init__.py
@@ -66,7 +66,7 @@ from ._shops import Shop, ShopStatus
 from ._supplied_products import SuppliedProduct
 from ._suppliers import Supplier, SupplierType
 from ._taxes import CustomerTaxGroup, Tax, TaxClass
-from ._units import SalesUnit
+from ._units import DisplayUnit, PiecesSalesUnit, SalesUnit, UnitInterface
 
 __all__ = [
     "AbstractOrderLine",
@@ -91,6 +91,7 @@ __all__ = [
     "CustomPaymentProcessor",
     "Currency",
     "DefaultOrderStatus",
+    "DisplayUnit",
     "FixedCostBehaviorComponent",
     "get_company_contact",
     "get_currency_precision",
@@ -116,6 +117,7 @@ __all__ = [
     "PaymentUrls",
     "PersistentCacheEntry",
     "PersonContact",
+    "PiecesSalesUnit",
     "PolymorphicShuupModel",
     "PolymorphicTranslatableShuupModel",
     "Product",
@@ -163,6 +165,7 @@ __all__ = [
     "Tax",
     "TaxClass",
     "TranslatableShuupModel",
+    "UnitInterface",
     "WaivingCostBehaviorComponent",
     "WeightBasedPriceRange",
     "WeightBasedPricingBehaviorComponent",