+++ shuup/core/order_creator/_source.py
@@ -28,6 +28,7 @@ from shuup.core.models import (
 )
 from shuup.core.pricing import Price, Priceful, TaxfulPrice, TaxlessPrice
 from shuup.core.taxing import should_calculate_taxes_automatically, TaxableItem
+from shuup.core.utils.line_unit_mixin import LineWithUnit
 from shuup.utils.decorators import non_reentrant
 from shuup.utils.i18n import format_money, is_existing_language
 from shuup.utils.money import Money
@@ -142,8 +143,7 @@ class OrderSource(object):
         self.zero_price = shop.create_price(0)
         self.create_price = self.zero_price.new
 
-        self._taxes_calculated = False
-        self._processed_lines_cache = None
+        self.uncache()  # Initialize caching variables
 
     def update(self, **values):
         for key, value in values.items():
@@ -368,11 +368,51 @@ class OrderSource(object):
     @property
     def product_count(self):
         """
-        Get the total number of products in this OrderSource.
+        Get the sum of product quantities in this order source.
 
-        :rtype: decimal.Decimal|int
+        Note: It is a bit silly to sum different units together.  Check
+        `smart_product_count` and `product_line_count` for other
+        options.
+
+        :rtype: decimal.Decimal
         """
         return sum([line.quantity for line in self.get_product_lines()])
+
+    @property
+    def smart_product_count(self):
+        """
+        Get the total number of separate products in this order source.
+
+        Quantities of lines, which have countable products, will be
+        summed and then number of lines with non-countable product units
+        will be added to that.  E.g. smart product count for a basket
+        containing 5 chocolate bars, 2 t-shirts and 2.5 kg of cocoa beans
+        would be 5 + 2 + 1 = 8.
+
+        Definition of "countable" here: If product has an unit that
+        allows presenting its quantities as a bare number (see
+        `~shuup.core.models.UnitInteface.allow_bare_number`) and its
+        quantity is an integral number, we assume that the unit is
+        similar to "Pieces" unit and those products being countable.
+        Other units are assumed to be non-countable.
+
+        :rtype: int
+        """
+        def count_in_line(line):
+            truncated_qty = int(line.quantity)
+            if not line.unit.allow_bare_number or truncated_qty != line.quantity:
+                return 1  # Non-countables or non-integral values counted as 1
+            return truncated_qty
+        return sum(count_in_line(line) for line in self.get_product_lines())
+
+    @property
+    def product_line_count(self):
+        """
+        Get the total number product lines in this order source
+
+        :rtype: int
+        """
+        return len(self.get_product_lines())
 
     def get_final_lines(self, with_taxes=False):
         """
@@ -427,6 +467,7 @@ class OrderSource(object):
         """
         self._processed_lines_cache = None
         self._taxes_calculated = False
+        self._object_cache = {}
 
     @non_reentrant
     def __compute_lines(self):
@@ -539,6 +580,23 @@ class OrderSource(object):
         product_lines = self.get_product_lines()
         return ((sum(l.product.gross_weight * l.quantity for l in product_lines)) if product_lines else 0)
 
+    def _get_object(self, model, pk):
+        """
+        Get model object from database by pk with caching.
+
+        Avoids same objects being loaded many times from the database
+        when constructing SourceLines in the same request.
+
+        :type model: type
+        :type pk: int|Any
+        :rtype: django.db.models.Model
+        """
+        obj = self._object_cache.get((model, pk))
+        if not obj:
+            obj = model.objects.get(pk=pk)
+            self._object_cache[(model, pk)] = obj
+        return obj
+
 
 def _collect_lines_from_signal(signal_results):
     for (receiver, response) in signal_results:
@@ -554,7 +612,7 @@ class LineSource(Enum):
     DISCOUNT_MODULE = 4
 
 
-class SourceLine(TaxableItem, Priceful):
+class SourceLine(TaxableItem, Priceful, LineWithUnit):
     """
     Line of OrderSource.
 
@@ -751,7 +809,7 @@ class SourceLine(TaxableItem, Priceful):
         for (name, model) in cls._OBJECT_FIELDS.items():
             id = result.pop(name + "_id", None)
             if id:
-                result[name] = model.objects.get(id=id)
+                result[name] = source._get_object(model, id)
 
         for name in cls._PRICE_FIELDS:
             value = result.get(name)