+++ shuup/front/static_src/js/update_price.js
@@ -15,7 +15,8 @@ window.updatePrice = function updatePrice(productId) {
     var data = {
         // In case productId is not available try to fallback to first input with correct name
         id: productId ? productId : $("input[name=product_id]").val(),
-        quantity: $quantity.val()
+        quantity: $quantity.val(),
+        unitType: $("#product-unit-type-" + productId).val()
     };
 
     var $simpleVariationSelect = $("#product-variations-" + productId);